chore: moved navigator hooks into navigator directory
