+++ app/component/itinerary/navigator/NaviContainer.js
@@ -4,7 +4,7 @@ import { legTime } from '../../../util/legUtils';
 import { itineraryShape, relayShape } from '../../../util/shapes';
 import NaviBottom from './NaviBottom';
 import NaviCardContainer from './NaviCardContainer';
-import { useRealtimeLegs } from '../hooks/useRealtimeLegs';
+import { useRealtimeLegs } from './hooks/useRealtimeLegs';
 
 function NaviContainer(
   { itinerary, focusToLeg, relayEnvironment, setNavigation, mapRef },
